Added a command handler, and added two new commands

+++ index.js
@@ -2,11 +2,24 @@ const Discord           = require("discord.js");
 const client            = new Discord.Client();
 const auth              = require("./config/auth.json");
 const statTracker       = require("./data/stattracker.js");
+const test              = require("./commands/test");
+const fs                = require("fs");
+const commandFiles      = fs.readdirSync('./commands').filter(file => file.endsWith('.js'));
 const chalk             = require("chalk");
+
 
 
 let digitCheck          = /\d+/g;
 let channels            = [];
+
+const cooldowns = new Discord.Collection();
+client.commands = new Discord.Collection();
+
+for (const file of commandFiles) {
+    const command = require(`./commands/${file}`);
+    client.commands.set(command.name, command);
+}
+
 
 client.login(auth.token);
 client.on ("ready", async () => {
@@ -21,66 +34,77 @@ client.on('guildCreate', async guild => {
     console.log(chalk.green("Server has been created, with guild id: " + guild.id));
 });
 
-// This all should probably be moved to it's own file for just commands.
 client.on("message", async message => {
-    if (message.author.bot) return;
-    let serverPrefix = await statTracker.getPrefix(message.guild.id);
-    const args = message.content.slice(serverPrefix.length).trim().split(/ +/g);
-    const command = args.shift().toLowerCase();
-    switch (command){
-        case "bindcount":
-            message.delete();
-            if (message.member.hasPermission("MANAGE_CHANNELS")){
-                await statTracker.updateCountChannel(message.guild.id, message.channel.id);
-                message.channel.send("Channel has been bound").then(msg => { msg.delete({ timeout: 120000})});
-                channels = await statTracker.getAllChannels();
-            } else 
-                message.channel.send("Sorry but you are missing the MANAGE_CHANNELS permission").then(msg => { msg.delete({ timeout: 120000 }) });
-            return;
-        break;
-        case 'setcount':
-            message.delete();
-            if (message.member.hasPermission("MANAGE_CHANNELS") || message.author.id == 167777692735766529){
-                let currentMessage = message.content;
-                if (!currentMessage.match(digitCheck) || currentMessage.includes(".") || currentMessage.includes(",")){
-                    message.channel.send("Sorry but that is not a valid value, please try again.").then(msg => { msg.delete({ timeout: 120000 }) });
-                } else {
-                    await statTracker.updateCount(message.guild.id, args[0]);
-                    message.channel.send("The current count for this server has been updated to. " + args[0]).then(msg => { msg.delete({ timeout: 120000 }) });
-                }
-            }
-            return;
-        break;
-        case 'setbreak':
-            message.delete();
-            if (message.member.hasPermission("MANAGE_CHANNELS")){
-                if (args.length < 0)
-                    message.channel.send("Sorry but you are missing an argument looking for true or false").then(msg => { msg.delete({ timeout: 30000 }) });
-                else if (args[0] == Boolean) {
-                    let temp = await statTracker.setContinuous(message.guild.id, args[0]);
-                    message.channel.send("Continuous message are now set to: " + temp).then(msg => { msg.delete({ timeout: 30000 }) });
-                }
-            } else {
-                message.channel.send("Sorry but you are missing the MANAGE_CHANNELS permission").then(msg => { msg.delete({ timeout: 120000 }) });
-            }
-            return;
-        break;
-        case 'getcount' :
-            message.delete();
-            let count = await statTracker.getCurrentCount(message.guild.id);
-            message.channel.send("The current count is... **" + count + "**").then(msg => { msg.delete({ timeout: 120000 }) });
-            return;
-        break;
-        case 'users' :
-            statTracker.testFunction(message.guild.id);
-            
-    }
-})
 
-client.on("message", async message => {
-    let serverPrefix = await statTracker.getPrefix(message.guild.id);
+    if (message.author.bot) return;
+    let serverPrefix;
+    
+    try {
+        serverPrefix = await statTracker.getPrefix(message.guild.id);
+    } catch (e) {
+        if (e instanceof TypeError)
+            serverPrefix = "!";
+        else    
+            console.error(e);
+    }
+    
+    let reply = "You did not provide any arguments.";
+
+    const args = message.content.slice(serverPrefix.length).trim().split(/ +/g);
+    const commandName = args.shift().toLowerCase();
+    const command = client.commands.get(commandName);
+    if (commandName == "test"){
+        await test.correctLeaderboard(message, message.guild.id);
+        return;
+    }
+
+    //Check if command exists
+    if (!client.commands.has(commandName)) return;
+
+    // Check for spam with cooldowns.
+    if (!cooldowns.has(command.name)) {
+        cooldowns.set(command.name, new Discord.Collection());
+    }
+    const now = Date.now();
+    const timestamps = cooldowns.get(command.name);
+    const cooldownAmount = (command.cooldown || 3) * 1000;
+
+    if (timestamps.has(message.author.id)) {
+        const expirationTime = timestamps.get(message.author.id) + cooldownAmount;
+
+        if (now < expirationTime) {
+            const timeLeft = (expirationTime - now) / 1000;
+            return message.reply(`please wait ${timeLeft.toFixed(1)} second(s), before using \`${command.name}\` again.`);
+        }
+    }
+
+    timestamps.set(message.author.id, now);
+    // resets the cooldown for user once the command time is up
+    setTimeout(() => timestamps.delete(message.author.id), cooldownAmount);
+
+    // Checking to see if the user added the required argument if not send an error
+    if (command.args && !args.length) {
+        if (command.usage)
+            reply += `\n The proper usage would be: ${serverPrefix}${command.name} ${command.usage}`;
+        return message.channel.send(reply).then(msg => { msg.delete({ timeout: 120000 }) });
+    }
+
+    if (commandName == "info"){
+        args[0] = await client.users.cache.get("167777692735766529").avatarURL();
+    }
+
+
+    // runs the command
+    try {
+        command.execute(message, args);
+        channels = await statTracker.getAllChannels();
+    } catch (err) {
+       return  console.error(err);
+
+    }
+    console.log ("Server prefix " + serverPrefix);
     let msg = message.content;
-    if (message.author.bot && !msg.includes(serverPrefix))
+    if (message.author.bot || msg.includes(serverPrefix) || message.channel.type === "dm")
         return;
     if (channels.includes(message.channel.id)){
         let currentMessage = message.content;
@@ -106,7 +130,7 @@ client.on("message", async message => {
 
 // Check to see if there is a message edited, if so move the counter down one and remove the message. 
 client.on("messageUpdate", async message => {
-    if (channels.includes(message.channel.id)){
+    if (channels.includes(message.channel.id) && !message.author.bot){
         let currentCount = await statTracker.getCurrentCount(message.guild.id);
         let lastUser = await statTracker.getLastUser(message.guild.id);
         message.delete();
